refactor: PRレビュー指摘の修正
- UNKNOWN_DATE_KEY と currentYearMonth() を ReceiptConstants.kt に一元化
- ReceiptViewModel の未使用コード (receiptsCache, getReceiptsForMonth) を削除
- DatabaseProvider のマイグレーション失敗時にログを追加
